requirements.txt
streamlit
pillow
numpy
pycryptodome